% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{basenameWithoutExt}
\alias{basenameWithoutExt}
\title{Get Base Name Without Extension}
\usage{
basenameWithoutExt(filename)
}
\arguments{
\item{filename}{Character. Path to the file.}
}
\value{
Character. File name without its extension.
}
\description{
Returns the base name of a file, removing its extension.
}
