% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bgr2rdensity}
\alias{bgr2rdensity}
\title{Convert BedGraph to RDensity Format}
\usage{
bgr2rdensity(inputfilename, outputfilename = NULL, binsize = 20)
}
\arguments{
\item{inputfilename}{Character. Input BedGraph file name.}

\item{outputfilename}{Character. Output RDensity file name.}

\item{binsize}{Integer. Bin size for density calculation (default: 20).}
}
\value{
None.
}
\description{
Converts a BedGraph file to RDensity format for density calculation.
}
\examples{
if (file.exists("input.bgr")) {
 bgr2rdensity("input.bgr", "output.rdensity", 20)
} else {
 cat("The file 'input.bgr' does not exist.\n")}
}
