% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{BuildCutCountProfileBAM}
\alias{BuildCutCountProfileBAM}
\title{Build CutCount Profile from BAM}
\usage{
BuildCutCountProfileBAM(
  bamfile = "",
  dataname = "",
  bgrfilename = "",
  ftablefile = "",
  refgenome = "hg19",
  chroms = NA,
  compress = T,
  atac = F
)
}
\arguments{
\item{bamfile}{Character. Path to BAM file.}

\item{dataname}{Character. Data name.}

\item{bgrfilename}{Character. Output BedGraph filename.}

\item{ftablefile}{Character. Frequency table file.}

\item{refgenome}{Character. Reference genome.}

\item{chroms}{Character vector. Chromosome names.}

\item{compress}{Logical. Compress output file (default: TRUE).}

\item{atac}{Logical. TRUE for ATAC-seq data (default: FALSE).}
}
\value{
None.
}
\description{
Builds a cut count profile from a BAM file and writes it to a BedGraph file.
}
