% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutCountProfile.R
\name{CalcExpectedRates}
\alias{CalcExpectedRates}
\title{Calculate Expected Cleavage Rates}
\usage{
CalcExpectedRates(tregion, dir = NA, nuccodes = NA, ftable = NA)
}
\arguments{
\item{tregion}{Data frame. Regions with chromosome, start, and end positions.}

\item{dir}{Data frame or vector. Direction information for each region.}

\item{nuccodes}{List. Nucleotide codes for each chromosome.}

\item{ftable}{Data frame. Cleavage rates table.}
}
\value{
Matrix of expected cleavage rates for each region.
}
\description{
Calculates expected cleavage rates for genomic regions based on nucleotide codes and a frequency table.
}
\examples{
# CalcExpectedRates(tregion, dir, nuccodes, ftable)
}
