% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFreqeuncyTableBAM.R
\name{calcFrequencyTableBAMChromosome}
\alias{calcFrequencyTableBAMChromosome}
\title{Calculate Frequency Table for a Chromosome from BAM File}
\usage{
calcFrequencyTableBAMChromosome(
  bamfile,
  chr,
  genome = NA,
  np = 6,
  mapdir = "",
  shifts = c(0, 0)
)
}
\arguments{
\item{bamfile}{Character. Path to the BAM file.}

\item{chr}{Character. Chromosome name.}

\item{genome}{BSgenome object. Reference genome.}

\item{np}{Integer. Length of the nucleotide sequence to consider (default: 6).}

\item{mapdir}{Character. Directory where mappability files are stored.}

\item{shifts}{Numeric vector. Shifts to apply (default: c(0,0)).}
}
\value{
Data frame. Frequency table for the chromosome.
}
\description{
Calculates a frequency table for a single chromosome from a BAM file.
}
\examples{
# calcFrequencyTableBAMChromosome("sample.bam", "chr1", genome, 6, "mapdir", c(0,0))
}
