% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{CalcObservedExpectedCuts}
\alias{CalcObservedExpectedCuts}
\title{Calculate Observed and Expected Cuts}
\usage{
CalcObservedExpectedCuts(
  S,
  motif,
  tdensity,
  motifcenter = 0,
  halfspan = 50,
  nuccodes = NA,
  motifwidth = 0,
  ftable = NA
)
}
\arguments{
\item{S}{AggregationPlotData object.}

\item{motif}{Data frame. Motif sites.}

\item{tdensity}{List. Cut count data per chromosome.}

\item{motifcenter}{Integer. Motif center position.}

\item{halfspan}{Integer. Window size around motif.}

\item{nuccodes}{List. Nucleotide codes.}

\item{motifwidth}{Integer. Motif width.}

\item{ftable}{Data frame. Frequency table.}
}
\value{
List with observed and expected cut counts.
}
\description{
Calculates observed and expected cut counts for motif sites.
}
