% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col2ColorCode}
\alias{col2ColorCode}
\title{Convert Color Name to RGB Code}
\usage{
col2ColorCode(col)
}
\arguments{
\item{col}{Character. Color name.}
}
\value{
Character. RGB color code as "R,G,B".
}
\description{
Converts a color name to an RGB string for use in track headers.
}
