% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combineTwoHotspots}
\alias{combineTwoHotspots}
\title{Combine Two Hotspot Files}
\usage{
combineTwoHotspots(csvfile1, csvfile2, name1, name2)
}
\arguments{
\item{csvfile1}{Character. Path to first hotspot file.}

\item{csvfile2}{Character. Path to second hotspot file.}

\item{name1}{Character. Name for first file.}

\item{name2}{Character. Name for second file.}
}
\value{
Character. Path to the pooled hotspot file.
}
\description{
Combines two hotspot (peak) files in CSV format by taking the union of hotspots.
}
