% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculBiasCorrectedCount.R
\name{convertBiasCorrectionTable}
\alias{convertBiasCorrectionTable}
\title{Convert Bias Correction Table}
\usage{
convertBiasCorrectionTable(ftablefile, np)
}
\arguments{
\item{ftablefile}{Character. Path to the frequency table file.}

\item{np}{Integer. Desired n-mer size.}
}
\value{
Data frame with converted bias correction table.
}
\description{
Converts a bias correction table to a different n-mer size and writes the result to a new file.
}
\examples{
# convertBiasCorrectionTable("Hexamer_FT_DS8497_mm9_withMap.txt", 4)
}
