% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{drawAggregatedObservedExpectedCutcountPlot}
\alias{drawAggregatedObservedExpectedCutcountPlot}
\title{Draw Aggregated Observed/Expected Cut Count Plot}
\usage{
drawAggregatedObservedExpectedCutcountPlot(
  mobs,
  mexp,
  motifcenter,
  halfwidth,
  title,
  motifwidth,
  linea,
  lineb,
  motiflogo,
  plot1yrange
)
}
\arguments{
\item{mobs}{Numeric vector. Mean observed cut counts.}

\item{mexp}{Numeric vector. Mean expected cut counts.}

\item{motifcenter}{Integer. Motif center position.}

\item{halfwidth}{Integer. Half-width for aggregation.}

\item{title}{Character. Plot title.}

\item{motifwidth}{Integer. Motif width.}

\item{linea}{Integer. Start of motif region.}

\item{lineb}{Integer. End of motif region.}

\item{motiflogo}{Character. Motif logo.}

\item{plot1yrange}{Numeric. Y-axis range for plot.}
}
\value{
None.
}
\description{
Draws a plot of aggregated observed and expected cut counts around motif sites.
}
