% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{filterHotspot}
\alias{filterHotspot}
\title{Filter Hotspot Data}
\usage{
filterHotspot(hotspot, category, threshold)
}
\arguments{
\item{hotspot}{Data frame. Hotspot data.}

\item{category}{Character. Category to filter by.}

\item{threshold}{Numeric. Threshold value.}
}
\value{
Data frame containing filtered hotspot data.
}
\description{
Filters hotspot data by category and threshold.
}
\examples{
#filterHotspot(hotspot, "MaxD", 5)
}
