% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagplot_masa.R
\name{gen_bagplot_masa}
\alias{gen_bagplot_masa}
\title{Generate a bagplot with optional HOMER results integration and point classification}
\usage{
gen_bagplot_masa(
  dat,
  dataname1 = dataname1,
  dataname2 = dataname2,
  qvaluethreshold_bagplot = 0.05,
  factor = 3,
  pdf = TRUE,
  html = FALSE,
  noBag = FALSE,
  noFence = FALSE,
  noMedianPoint = FALSE,
  pdfwidth = 16,
  pdfheight = 12,
  textsizefactor = 1,
  printmode = FALSE,
  include_homer_results = FALSE,
  homer_results_file = NULL,
  pval_threshold_homer = 0.05,
  use_cluster_motif_db = TRUE
)
}
\arguments{
\item{dat}{Data frame containing bagplot input data.}

\item{dataname1}{Name of the first dataset (e.g., control).}

\item{dataname2}{Name of the second dataset (e.g., treatment).}

\item{qvaluethreshold_bagplot}{Threshold for q-value significance (default: 0.05).}

\item{factor}{Factor for bagplot scaling (default: 3).}

\item{pdf}{Whether to save the plot as a PDF (default: TRUE).}

\item{include_homer_results}{Whether to include HOMER results (default: FALSE).}

\item{homer_results_file}{Path to the HOMER results file (e.g., knownResults.txt).}

\item{pval_threshold_homer}{P-value threshold for HOMER results (default: 0.05).}

\item{use_cluster_motif_db}{Whether to apply motif name extraction specific to the cluster_motif database (default: FALSE).}
}
\value{
A ggplot object with the bagplot.
}
\description{
Generate a bagplot with optional HOMER results integration and point classification
}
