% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFreqeuncyTableBAM.R
\name{getNucleotideString}
\alias{getNucleotideString}
\title{Get Nucleotide String Combinations}
\usage{
getNucleotideString(np)
}
\arguments{
\item{np}{Integer. Length of the nucleotide strings to generate.}
}
\value{
Character vector of all possible nucleotide strings of length np.
}
\description{
Generates all possible nucleotide strings of a given length.
}
\examples{
getNucleotideString(2)
getNucleotideString(3)
}
