% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getOutputDirAndFilename}
\alias{getOutputDirAndFilename}
\title{Get Output Directory and Filename}
\usage{
getOutputDirAndFilename(outputdir_base, dataname, sitefilename, halfwidth)
}
\arguments{
\item{outputdir_base}{Character. Base output directory.}

\item{dataname}{Character. Data name.}

\item{sitefilename}{Character. Site file name.}

\item{halfwidth}{Integer. Half-width value.}
}
\value{
List with elements 'dir' and 'filename'.
}
\description{
Constructs output directory and filename for results based on input parameters.
}
