% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{GFoot-class}
\alias{GFoot-class}
\alias{GFoot}
\title{GFoot Class}
\value{
A GFoot object.
}
\description{
S4 class to store all data and options for a GFoot analysis.
}
\section{Slots}{

\describe{
\item{\code{control}}{CutCount. Control cut count data.}

\item{\code{treatment}}{CutCount. Treatment cut count data.}

\item{\code{sitefile}}{Character. Path to site file.}

\item{\code{motifDB}}{MotifDB. Motif database object.}

\item{\code{gfootoption}}{GFootOption. GFoot analysis options.}

\item{\code{outputdir}}{Character. Output directory.}

\item{\code{cachedir}}{Character. Cache directory.}

\item{\code{outputfiles}}{Character. Output files generated.}
}}

\examples{
#gfoot <- GFoot(control = cc1, treatment = cc2, sitefile = "sites.txt", motifDB = mdb, gfootoption = opt)
}
