% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{GFootOption-class}
\alias{GFootOption-class}
\alias{GFootOption}
\title{GFootOption Class}
\value{
A GFootOption object.
}
\description{
S4 class to store GFoot analysis options.
}
\section{Slots}{

\describe{
\item{\code{biasfile}}{Character. Path to bias correction file.}

\item{\code{hexamer_pattern}}{Character. Pattern for hexamer code files.}
}}

