% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_chip.R
\name{intersect_atac_with_chip}
\alias{intersect_atac_with_chip}
\title{Run bedtoolsr intersect on peak files with reference using bedtoolsr}
\usage{
intersect_atac_with_chip(
  atac_dir,
  chip_file,
  results_dir,
  result_name,
  file_extension = "narrowPeak"
)
}
\arguments{
\item{atac_dir}{Path to the directory containing peak files (e.g., ATAC-seq)}

\item{chip_file}{Path to the reference chip-seq BED file}

\item{results_dir}{Path to the directory where result \code{.bed} files will be stored}

\item{result_name}{Suffix used in the output file names}

\item{file_extension}{File extension to match (e.g., "bed", "narrowPeak"). Do not include the dot.}
}
\value{
This function does not return an R object. It writes \code{.bed} files to the specified output directory.

Use -wa and -u as intersect option to keep only the unique ATAc-seq peaks intersected with the chip-seq peaks
}
\description{
This function runs \code{bedtoolsr::bt.intersect} for each peak file in a specified directory
against a reference BED file (e.g., H3K27ac ChIP-seq peaks). Each result is saved in the
output directory using a customizable result name suffix.
}
