% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculBiasCorrectedCount.R
\name{intersect_intervals_fast_parallel_index2}
\alias{intersect_intervals_fast_parallel_index2}
\title{Fast Parallel Interval Intersection}
\usage{
intersect_intervals_fast_parallel_index2(P)
}
\arguments{
\item{P}{List. Contains Ast, Aed, Bst, Bed vectors.}
}
\value{
List of intersection indices.
}
\description{
Finds intersections between two sets of intervals in parallel.
}
\examples{
# intersect_intervals_fast_parallel_index2(list(Ast = ..., Aed = ..., Bst = ..., Bed = ...))
}
