% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutCountProfile.R
\name{loadChromosomeRange}
\alias{loadChromosomeRange}
\title{Load Chromosome Ranges for Reference Genome}
\usage{
loadChromosomeRange(refgenome)
}
\arguments{
\item{refgenome}{Character. Reference genome name (e.g., "mm10", "mm9", "hg19", "hg18", "mm8","hg38").}
}
\value{
Named list of chromosome ranges (start and end positions).
}
\description{
Returns a list of chromosome ranges for a specified reference genome.
}
\examples{
#loadChromosomeRange("mm10")
}
