% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{loadCutcount}
\alias{loadCutcount}
\alias{loadCutcount,GFoot-method}
\title{Load Cut Count Data for GFoot Object (S4 Generic)}
\usage{
loadCutcount(obj)

\S4method{loadCutcount}{GFoot}(obj)
}
\arguments{
\item{obj}{S4 object of class GFoot.}
}
\value{
S4 object of class GFoot with cutcount slots populated.
}
\description{
Loads cut count data for control and treatment in a GFoot object.
}
\section{Functions}{
\itemize{
\item \code{loadCutcount(GFoot)}: Method for GFoot objects.

}}
\examples{
# loadCutcount(gfoot_obj)
}
