% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFreqeuncyTableBAM.R
\name{loadReferenceGenome}
\alias{loadReferenceGenome}
\title{Load Reference Genome}
\usage{
loadReferenceGenome(refgenome)
}
\arguments{
\item{refgenome}{Character. Reference genome name ("mm9", "mm10", "hg19","hg38").}
}
\value{
BSgenome object representing the reference genome.
}
\description{
Loads a reference genome as a BSgenome object.
}
\examples{
# loadReferenceGenome("mm10")
}
