% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MakeBiasCorrectionTableBAM}
\alias{MakeBiasCorrectionTableBAM}
\title{Generate n-mer Bias Tables from BAM}
\usage{
MakeBiasCorrectionTableBAM(
  bamfile = "",
  outfile = "",
  refgenome = "",
  np = 6,
  mapdir = "",
  atac = F
)
}
\arguments{
\item{bamfile}{Character. Path to the input BAM file.}

\item{outfile}{Character. Output filename.}

\item{refgenome}{Character. Reference genome.}

\item{np}{Integer. Number of base pairs (default: 6).}

\item{mapdir}{Character. Directory to mappability files.}

\item{atac}{Logical. TRUE for ATAC-seq data.}
}
\value{
Character. Output filename.
}
\description{
Generates n-mer bias tables from a BAM file, optionally using mappability files.
}
