% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeCutCountBAM}
\alias{makeCutCountBAM}
\title{Make CutCount BedGraph File from BAM}
\usage{
makeCutCountBAM(bamFile, refgenome = "mm10")
}
\arguments{
\item{refgenome}{Character. Reference genome (e.g., "mm9", "hg19").}

\item{bamfile}{Character. Path to the input BAM file.}
}
\value{
Character. BedGraph filename.
}
\description{
Creates a BedGraph file containing cut counts from a BAM file.
}
