% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeCutCountBAMWithName}
\alias{makeCutCountBAMWithName}
\title{Make CutCount BedGraph File from BAM with Name}
\usage{
makeCutCountBAMWithName(bamFile, name = "", refgenome = "mm10", atac = F)
}
\arguments{
\item{bamFile}{Character. Path to the input BAM file.}

\item{name}{Character. Sample name.}

\item{refgenome}{Character. Reference genome (default: "mm10").}

\item{atac}{Logical. TRUE for ATAC-seq data (default: FALSE).}
}
\value{
Character. Path to the compressed BedGraph file (.bgr.gz).
}
\description{
Creates a BedGraph file containing cut counts from a BAM file, using a sample name.
}
