% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{MCCORES}
\alias{MCCORES}
\title{All the classes that we need to use in the package}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
MCCORES
}
\value{
nothing
}
\description{
All the classes that we need to use in the package
}
\section{Slots}{

\describe{
\item{\code{file}}{character.}

\item{\code{cutcount}}{list.}

\item{\code{count}}{numeric.}

\item{\code{name}}{character.}
}}

\keyword{datasets}
