% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{MotifDB-class}
\alias{MotifDB-class}
\alias{MotifDB}
\title{MotifDB Class}
\value{
A MotifDB object.
}
\description{
S4 class to store motif database information.
}
\section{Slots}{

\describe{
\item{\code{motiflistfile}}{Character. Path to motif list file.}

\item{\code{directory}}{Character. Directory containing motif files.}
}}

\examples{
# mdb <- MotifDB(motiflistfile = "motifs.txt", directory = "motif_dir")
}
