% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{MultipleAggregationPlot}
\alias{MultipleAggregationPlot}
\title{Multiple Aggregation Plot}
\usage{
MultipleAggregationPlot(
  dataname1 = "",
  dataname2 = "",
  motiflistfile1 = "",
  motiflistfile2 = "",
  outputfile1 = "",
  outputfile2 = "",
  title = "",
  threshold = 0,
  datadir1 = "",
  datadir2 = "",
  outputdir = " ",
  scalefactor = c(1, 1),
  range = NA,
  logyrange = 0,
  graphout = F
)
}
\arguments{
\item{dataname1}{Character. Name of first dataset.}

\item{dataname2}{Character. Name of second dataset.}

\item{motiflistfile1}{Character. Path to first motif list file.}

\item{motiflistfile2}{Character. Path to second motif list file.}

\item{outputfile1}{Character. Path to first output data file.}

\item{outputfile2}{Character. Path to second output data file.}

\item{title}{Character. Plot title.}

\item{threshold}{Numeric. Threshold for filtering.}

\item{datadir1}{Character. Directory for first dataset.}

\item{datadir2}{Character. Directory for second dataset.}

\item{outputdir}{Character. Output directory.}

\item{scalefactor}{Numeric vector. Scale factors for data.}

\item{range}{Optional. Range of motifs to analyze.}

\item{logyrange}{Numeric vector. Y-axis range for log ratio plot.}

\item{graphout}{Logical. Output graphs.}
}
\value{
None.
}
\description{
Draws multiple aggregation plots for motif comparisons and saves results.
}
