% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pool_hotspots3}
\alias{pool_hotspots3}
\title{Pool Hotspots from Multiple Files}
\usage{
pool_hotspots3(hotspotdata)
}
\arguments{
\item{hotspotdata}{List of data frames. Hotspot data from multiple files.}
}
\value{
Data frame. Merged hotspot data.
}
\description{
Merges two or more MACS2 peak files into a single hotspot file.
}
