% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printf}
\alias{printf}
\title{Print Formatted Output}
\usage{
printf(...)
}
\arguments{
\item{...}{Arguments passed to sprintf.}
}
\value{
None.
}
\description{
Prints formatted output to the console, similar to printf in C.
}
\examples{
#printf("Hello \%s\n", "world")
}
