% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{readannot}
\alias{readannot}
\title{Read Annotation File}
\usage{
readannot(filename)
}
\arguments{
\item{filename}{Character. Path to the annotation file.}
}
\value{
Data frame containing annotation data.
}
\description{
Reads an annotation file in CSV format.
}
\examples{
#readannot("annotation.csv")
}
