% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readBAMIndex}
\alias{readBAMIndex}
\title{Read BAM Index}
\usage{
readBAMIndex(bamFile, refgenome = "mm9")
}
\arguments{
\item{bamFile}{Character. Path to BAM file.}

\item{refgenome}{Character. Reference genome ("mm9", "mm10", "hg19","hg38).}
}
\value{
Data frame with columns chr, maxloc, num, etc.
}
\description{
Reads BAM index statistics and returns a data frame for chromosomes in the reference genome.
}
