% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readbed}
\alias{readbed}
\title{Read BED File}
\usage{
readbed(filename)
}
\arguments{
\item{filename}{Character. Path to BED file.}
}
\value{
Data frame with columns chr, st, ed.
}
\description{
Reads a BED file and returns a data frame with columns chr, st, ed.
}
