% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{readbed_as_csv}
\alias{readbed_as_csv}
\title{Read BED File as CSV}
\usage{
readbed_as_csv(filename)
}
\arguments{
\item{filename}{Character. Path to the BED file.}
}
\value{
Data frame with columns "chr", "st", "ed", "MaxD", "AveD", "Zscore".
}
\description{
Reads a BED file and returns a data frame with additional columns.
}
\examples{
#readbed_as_csv("bedfile.bed")
}
