% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R, R/utils.R
\name{readbgr}
\alias{readbgr}
\title{Read BedGraph File}
\usage{
readbgr(filename)

readbgr(filename)
}
\arguments{
\item{filename}{Character. Path to the BedGraph file.}
}
\value{
Data frame with columns "chr", "st", "ed", "value".

Data frame with columns chr, st, ed, value.
}
\description{
Reads a BedGraph file and returns a data frame with columns "chr", "st", "ed", and "value".

Reads a BedGraph file and returns a data frame with columns chr, st, ed, and value.
}
\examples{
#readbgr("example.bed")
}
