% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readcutcount2}
\alias{readcutcount2}
\title{Read Cut Count Data from BedGraph}
\usage{
readcutcount2(cutcountfile)
}
\arguments{
\item{cutcountfile}{Character. Path to the BedGraph file.}
}
\value{
List of data frames, one per chromosome.
}
\description{
Reads cut count data from a BedGraph file and splits it by chromosome.
}
