% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{readCutCountOnSites}
\alias{readCutCountOnSites}
\title{Read Cut Count on Sites}
\usage{
readCutCountOnSites(cutcount_file = "", site_file = "")
}
\arguments{
\item{cutcount_file}{Character. Path to cut count file.}

\item{site_file}{Character. Path to site file.}
}
\value{
List. Reduced cut count data per chromosome.
}
\description{
Reads and reduces cut count data based on overlapping genomic sites.
}
\examples{
# readCutCountOnSites("cutcount.bgr", "sites.txt")
}
