% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.R
\name{readCutCountSites}
\alias{readCutCountSites}
\alias{readCutCountSites,CutCount-method}
\title{Read Cut Count Sites (S4 Generic)}
\usage{
readCutCountSites(obj, site_filename)

\S4method{readCutCountSites}{CutCount}(obj, site_filename)
}
\arguments{
\item{obj}{S4 object of class CutCount.}

\item{site_filename}{Character. Path to the site file.}
}
\value{
S4 object of class CutCount with cutcount slot populated.
}
\description{
Generic method to read cut count data for specified sites.
}
\section{Functions}{
\itemize{
\item \code{readCutCountSites(CutCount)}: Method for CutCount objects.

}}
\examples{
# readCutCountSites(obj, "sites.txt")
}
