% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readGenome}
\alias{readGenome}
\title{Read Reference Genome}
\usage{
readGenome(refgenome)
}
\arguments{
\item{refgenome}{Character. Reference genome name.}
}
\value{
BSgenome object.
}
\description{
Loads a reference genome as a BSgenome object.
}
