% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_function.R
\name{readhotspot}
\alias{readhotspot}
\title{Read Hotspot File}
\usage{
readhotspot(filename)
}
\arguments{
\item{filename}{Character. Path to the hotspot file.}
}
\value{
Data frame containing hotspot data.
}
\description{
Reads a hotspot file in CSV format.
}
\examples{
#readhotspot("hotspot.csv")
}
