% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutCountProfile.R
\name{readNucleotideCodeForChromosomeForCuts}
\alias{readNucleotideCodeForChromosomeForCuts}
\title{Read Nucleotide Codes at Cut Sites}
\usage{
readNucleotideCodeForChromosomeForCuts(nuccodefile, np)
}
\arguments{
\item{nuccodefile}{Character. Path to nucleotide code RDS file.}

\item{np}{Integer. n-mer size.}
}
\value{
Integer vector of nucleotide codes at cut sites.
}
\description{
Reads nucleotide codes from a file and returns codes at cut sites, shifted by half the n-mer size.
}
\examples{
# readNucleotideCodeForChromosomeForCuts("nuccode_chr1_4mer.dat", 4)
}
