% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R, R/method.R
\name{run}
\alias{run}
\alias{run,GFoot-method}
\title{Run GFoot Analysis (S4 Generic)}
\usage{
run(
  obj,
  range = NA,
  graphout = F,
  yrange = c(-2.2, 0.5),
  mc.cores = MCCORES,
  run_set = NA
)

\S4method{run}{GFoot}(
  obj,
  range = NA,
  graphout = F,
  yrange = c(-2.2, 0.5),
  mc.cores = MCCORES,
  run_set = NA
)

run(
  obj,
  range = NA,
  graphout = F,
  yrange = c(-2.2, 0.5),
  mc.cores = MCCORES,
  run_set = NA
)

\S4method{run}{GFoot}(
  obj,
  range = NA,
  graphout = F,
  yrange = c(-2.2, 0.5),
  mc.cores = MCCORES,
  run_set = NA
)
}
\arguments{
\item{obj}{S4 object of class GFoot.}

\item{range}{Optional. Range for analysis.}

\item{graphout}{Logical. If TRUE, output graphs.}

\item{yrange}{Numeric vector. Y-axis range for plots.}

\item{mc.cores}{Integer. Number of CPU cores to use.}

\item{run_set}{Optional. Additional run set parameter.}
}
\value{
GFoot object with output files.

S4 object of class GFoot with output files.
}
\description{
Generic method to run GFoot analysis.

Runs GFoot analysis for a GFoot object.
}
\section{Functions}{
\itemize{
\item \code{run(GFoot)}: Method for GFoot objects.

\item \code{run(GFoot)}: Method for GFoot objects.

}}
\examples{
# run(gfoot_obj, range = 1:10, graphout = TRUE)
}
