% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeBGRheader}
\alias{writeBGRheader}
\title{Write BedGraph Header}
\usage{
writeBGRheader(filename, name, desc, option, color)
}
\arguments{
\item{filename}{Character. Path to the BedGraph file.}

\item{name}{Character. Track name.}

\item{desc}{Character. Track description.}

\item{option}{Character. Track options.}

\item{color}{Character. Track color code.}
}
\value{
None.
}
\description{
Writes a header line to a BedGraph file.
}
